Move epilogue fragment into login sub package
